fix: prevent --outtype KeyError, duplicate log lines, and np.prod overflow
Three small correctness fixes surfaced by the optimization scan:

1. safetensors_to_gguf.py: --outtype accepted k-quant choices (q4_k, q5_0,
   ...) that were absent from ftype_map, so selecting one crashed with a
   KeyError before conversion. Restrict the choices to the formats the
   converter actually supports (f32/f16/bf16/q8_0/tq1_0/tq2_0/auto) and point
   users to quantize_gguf.py for k-quants.

2. safetensors_to_gguf.py & quantize_gguf.py: a module-level StreamHandler was
   attached to the named logger *and* logging.basicConfig() was called in
   main(), so every log line printed twice. Drop the module-level handler and
   rely on basicConfig.

3. analyze_gguf.py: np.prod(shape) can overflow the platform int for large
   tensors, yielding wrong sizes. Compute with dtype=np.int64.

+++ analyze_gguf.py
@@ -81,7 +81,8 @@ def analyze_gguf_model(model_path):
             tensor_name = tensor.name
             tensor_type = tensor.tensor_type
             tensor_shape = tensor.shape
-            tensor_size = np.prod(tensor_shape) * gguf.get_type_size(tensor_type)
+            # Use int64 to avoid overflow when multiplying large tensor dimensions
+            tensor_size = int(np.prod(tensor_shape, dtype=np.int64)) * gguf.get_type_size(tensor_type)
             total_size += tensor_size
             
             # Update tensor type statistics
